+++ gmm_net/models/gmm_net_for_own_opp_team_performance.py
